+++ internal/push-notification/domain/model/message.go
