+++ internal/push-notification/domain/repository/fcm.go
@@ -3,7 +3,7 @@ package repository
 import (
 	"context"
 
-	"github.com/abyssparanoia/rapid-go/push-notification/domain/model"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/domain/model"
 )
 
 // Fcm ... fcm repository interface
