+++ internal/push-notification/domain/repository/token.go
@@ -1,7 +1,7 @@
 package repository
 
 import (
-	"github.com/abyssparanoia/rapid-go/push-notification/domain/model"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/domain/model"
 
 	"context"
 )
