+++ internal/push-notification/domain/service/token.go
@@ -1,7 +1,7 @@
 package service
 
 import (
-	"github.com/abyssparanoia/rapid-go/push-notification/domain/model"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/domain/model"
 
 	"context"
 )
