+++ internal/push-notification/domain/service/token_impl.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	"github.com/abyssparanoia/rapid-go/internal/pkg/log"
-	"github.com/abyssparanoia/rapid-go/push-notification/domain/model"
-	"github.com/abyssparanoia/rapid-go/push-notification/domain/repository"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/domain/model"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/domain/repository"
 )
 
 type token struct {
