+++ internal/push-notification/handler/api/message.go
@@ -6,9 +6,9 @@ import (
 	"github.com/abyssparanoia/rapid-go/internal/pkg/errcode"
 	"github.com/abyssparanoia/rapid-go/internal/pkg/parameter"
 	"github.com/abyssparanoia/rapid-go/internal/pkg/renderer"
-	"github.com/abyssparanoia/rapid-go/push-notification/usecase/input"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/usecase/input"
 
-	"github.com/abyssparanoia/rapid-go/push-notification/usecase"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/usecase"
 	validator "gopkg.in/go-playground/validator.v9"
 )
