+++ internal/push-notification/infrastructure/internal/entity/message.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"firebase.google.com/go/messaging"
-	"github.com/abyssparanoia/rapid-go/push-notification/domain/model"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/domain/model"
 )
 
 // NewMessageFromModel ... new message from model
