+++ internal/push-notification/infrastructure/internal/entity/token.go
@@ -3,7 +3,7 @@ package entity
 import (
 	"cloud.google.com/go/firestore"
 
-	"github.com/abyssparanoia/rapid-go/push-notification/domain/model"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/domain/model"
 )
 
 // Token ... token entity
