+++ internal/push-notification/infrastructure/repository/token_impl.go
@@ -6,9 +6,9 @@ import (
 	"cloud.google.com/go/firestore"
 	"github.com/abyssparanoia/rapid-go/internal/pkg/gluefirestore"
 	"github.com/abyssparanoia/rapid-go/internal/pkg/log"
-	"github.com/abyssparanoia/rapid-go/push-notification/domain/model"
-	"github.com/abyssparanoia/rapid-go/push-notification/domain/repository"
-	"github.com/abyssparanoia/rapid-go/push-notification/infrastructure/internal/entity"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/domain/model"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/domain/repository"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/infrastructure/internal/entity"
 )
 
 type token struct {
