+++ internal/push-notification/usecase/token_impl.go
@@ -4,11 +4,11 @@ import (
 	"context"
 
 	"github.com/abyssparanoia/rapid-go/internal/pkg/log"
-	"github.com/abyssparanoia/rapid-go/push-notification/config"
-	"github.com/abyssparanoia/rapid-go/push-notification/domain/model"
-	"github.com/abyssparanoia/rapid-go/push-notification/domain/repository"
-	"github.com/abyssparanoia/rapid-go/push-notification/domain/service"
-	"github.com/abyssparanoia/rapid-go/push-notification/usecase/input"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/config"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/domain/model"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/domain/repository"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/domain/service"
+	"github.com/abyssparanoia/rapid-go/internal/push-notification/usecase/input"
 )
 
 type token struct {
